Correção do namespace Pitangueiros para Pitang
