+++ Pitang.Kifome.Application.Services/Pitang.Kifome.Application.Services.csproj
@@ -59,10 +59,6 @@
       <Project>{48865435-4775-462e-b1e2-a4f028b0e2de}</Project>
       <Name>Pitang.Kifome.Domain.Contracts</Name>
     </ProjectReference>
-    <ProjectReference Include="..\Pitangueiros.Kifome.Crosscuting.IoC\Pitangueiros.Kifome.Crosscuting.IoC.csproj">
-      <Project>{254f1376-9676-4447-8bb1-65ecc3bacf89}</Project>
-      <Name>Pitangueiros.Kifome.Crosscuting.IoC</Name>
-    </ProjectReference>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
