+++ Pitang.Kifome.Crosscuting.IoC/Pitang.Kifome.Crosscuting.IoC.csproj
@@ -4,11 +4,11 @@
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{76CA5D59-E522-4A0F-BEF8-2AE0E8AFD26C}</ProjectGuid>
+    <ProjectGuid>{254F1376-9676-4447-8BB1-65ECC3BACF89}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>Pitangueiros.Kifome.Infra.Repositories</RootNamespace>
-    <AssemblyName>Pitangueiros.Kifome.Infra.Repositories</AssemblyName>
+    <RootNamespace>Pitang.Kifome.Crosscuting.IoC</RootNamespace>
+    <AssemblyName>Pitang.Kifome.Crosscuting.IoC</AssemblyName>
     <TargetFrameworkVersion>v4.6.1</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <TargetFrameworkProfile />
@@ -34,14 +34,13 @@
     <OutputPath>bin\joao\</OutputPath>
   </PropertyGroup>
   <ItemGroup>
-    <Reference Include="EntityFramework, Version=6.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089, processorArchitecture=MSIL">
-      <HintPath>..\packages\EntityFramework.6.2.0\lib\net45\EntityFramework.dll</HintPath>
+    <Reference Include="Castle.Core, Version=4.0.0.0, Culture=neutral, PublicKeyToken=407dd0808d44fbdc, processorArchitecture=MSIL">
+      <HintPath>..\packages\Castle.Core.4.2.1\lib\net45\Castle.Core.dll</HintPath>
     </Reference>
-    <Reference Include="EntityFramework.SqlServer, Version=6.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089, processorArchitecture=MSIL">
-      <HintPath>..\packages\EntityFramework.6.2.0\lib\net45\EntityFramework.SqlServer.dll</HintPath>
+    <Reference Include="Castle.Windsor, Version=4.0.0.0, Culture=neutral, PublicKeyToken=407dd0808d44fbdc, processorArchitecture=MSIL">
+      <HintPath>..\packages\Castle.Windsor.4.1.0\lib\net45\Castle.Windsor.dll</HintPath>
     </Reference>
     <Reference Include="System" />
-    <Reference Include="System.ComponentModel.DataAnnotations" />
     <Reference Include="System.Configuration" />
     <Reference Include="System.Core" />
     <Reference Include="System.Runtime.Remoting" />
@@ -54,35 +53,11 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="IocManager.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Repository\CommentRepository.cs" />
-    <Compile Include="Repository\Context.cs" />
-    <Compile Include="Repository\CustomerRepository.cs" />
-    <Compile Include="Repository\DeliveryRepository.cs" />
-    <Compile Include="Repository\GarnishRepository.cs" />
-    <Compile Include="Repository\MealRepository.cs" />
-    <Compile Include="Repository\MenuRepository.cs" />
-    <Compile Include="Repository\OrderRepository.cs" />
-    <Compile Include="Repository\SellerRepository.cs" />
-    <Compile Include="Repository\UserRepository.cs" />
   </ItemGroup>
   <ItemGroup>
-    <None Include="App.config" />
     <None Include="packages.config" />
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\Pitang.Kifome.Domain.Contracts\Pitang.Kifome.Domain.Contracts.csproj">
-      <Project>{48865435-4775-462e-b1e2-a4f028b0e2de}</Project>
-      <Name>Pitang.Kifome.Domain.Contracts</Name>
-    </ProjectReference>
-    <ProjectReference Include="..\Pitang.Kifome.Domain.Entities\Pitang.Kifome.Domain.Entities.csproj">
-      <Project>{ec121a18-ffca-435e-85dc-00c2ffb9d8ed}</Project>
-      <Name>Pitang.Kifome.Domain.Entities</Name>
-    </ProjectReference>
-    <ProjectReference Include="..\Pitangueiros.Kifome.Crosscuting.IoC\Pitangueiros.Kifome.Crosscuting.IoC.csproj">
-      <Project>{254f1376-9676-4447-8bb1-65ecc3bacf89}</Project>
-      <Name>Pitangueiros.Kifome.Crosscuting.IoC</Name>
-    </ProjectReference>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
