+++ Pitang.Kifome.Crosscuting.IoC/Properties/AssemblyInfo.cs
@@ -5,11 +5,11 @@ using System.Runtime.InteropServices;
 // As informações gerais sobre um assembly são controladas por
 // conjunto de atributos. Altere estes valores de atributo para modificar as informações
 // associada a um assembly.
-[assembly: AssemblyTitle("Pitangueiros.Kifome.Crosscuting.IoC")]
+[assembly: AssemblyTitle("Pitang.Kifome.Crosscuting.IoC")]
 [assembly: AssemblyDescription("")]
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("")]
-[assembly: AssemblyProduct("Pitangueiros.Kifome.Crosscuting.IoC")]
+[assembly: AssemblyProduct("Pitang.Kifome.Crosscuting.IoC")]
 [assembly: AssemblyCopyright("Copyright ©  2017")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
