+++ Pitang.Kifome.Domain.Services.IoC/Pitang.Kifome.Domain.Services.IoC.csproj
@@ -68,10 +68,6 @@
       <Project>{32cc437f-1296-476d-b15a-fb4bb6d0b49f}</Project>
       <Name>Pitang.Kifome.Domain.Services</Name>
     </ProjectReference>
-    <ProjectReference Include="..\Pitangueiros.Kifome.Crosscuting.IoC\Pitangueiros.Kifome.Crosscuting.IoC.csproj">
-      <Project>{254f1376-9676-4447-8bb1-65ecc3bacf89}</Project>
-      <Name>Pitangueiros.Kifome.Crosscuting.IoC</Name>
-    </ProjectReference>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
