+++ Pitang.Kifome.Infra.Repositories/Repository/Context.cs
@@ -1,12 +1,6 @@
 ﻿using Pitang.Kifome.Domain.Entities;
-using System;
-using System.Collections.Generic;
-using System.Data.Entity;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace Pitangueiros.Kifome.Infra.Repositories.Repository
+namespace Pitang.Kifome.Infra.Repositories.Repository
 {
     public class Context : DbContext
     {
