+++ Pitang.Kifome.Infra.Repositories/Repository/CustomerRepository.cs
@@ -1,12 +1,9 @@
-﻿using Pitang.Kifome.Domain.Contracts.Repositories;
-using Pitang.Kifome.Domain.Entities;
-using System;
+﻿using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+using Pitang.Kifome.Domain.Contracts.Repositories;
+using Pitang.Kifome.Domain.Entities;
 
-namespace Pitangueiros.Kifome.Infra.Repositories.Repository
+namespace Pitang.Kifome.Infra.Repositories.Repository
 {
     public class CustomerRepository : ICustomerRepository
     {
