+++ Pitang.Kifome.Infra.Repositories/Repository/SellerRepository.cs
@@ -1,41 +1,38 @@
-﻿using Pitang.Kifome.Domain.Contracts.Repositories;
-using Pitang.Kifome.Domain.Entities;
-using System;
+﻿using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+using Pitang.Kifome.Domain.Contracts.Repositories;
+using Pitang.Kifome.Domain.Entities;
 
-namespace Pitangueiros.Kifome.Infra.Repositories.Repository
+namespace Pitang.Kifome.Infra.Repositories.Repository
 {
-    public class OrderRepository : IOrderRepository
+    public class SellerRepository : ISellerRepository
     {
         public void Delete(int id)
         {
             throw new NotImplementedException();
         }
 
-        public void Insert(Order entity)
+        public void Insert(Seller entity)
         {
             throw new NotImplementedException();
         }
 
-        public Order Select(int id)
+        public void InsertMenu(List<Meal> meals, DateTime initialHour, DateTime finalHour, int limitOfMeals)
         {
             throw new NotImplementedException();
         }
 
-        public List<Order> SelectAll()
+        public Seller Select(int id)
         {
             throw new NotImplementedException();
         }
 
-        public List<Order> SelectAllByUser(User user)
+        public List<Seller> SelectAll()
         {
             throw new NotImplementedException();
         }
 
-        public void Update(Order entity)
+        public void Update(Seller entity)
         {
             throw new NotImplementedException();
         }
